backend/requirements.txt
flask==2.3.3
numpy==1.26.4
scikit-learn==1.4.1.post1
python-dotenv==1.0.1
requests==2.31.0
pymongo==4.6.1
PyJWT==2.8.0
google-generativeai==0.3.1
flask-cors==4.0.0 